+++ packages/compiler/src/core/program.ts
@@ -1,5 +1,4 @@
 import { EmitterOptions } from "../config/types.js";
-import { createAssetEmitter } from "../emitter-framework/asset-emitter.js";
 import { validateEncodedNamesConflicts } from "../lib/encoded-names.js";
 import { MANIFEST } from "../manifest.js";
 import {
@@ -37,6 +36,7 @@ import { isImportStatement, parse, parseStandaloneTypeReference } from "./parser
 import { getDirectoryPath, joinPaths, resolvePath } from "./path-utils.js";
 import { createProjector } from "./projector.js";
 import { createSourceFile } from "./source-file.js";
+import { StateMap, StateSet, createStateAccessors } from "./state-accessors.js";
 import {
   CompilerHost,
   Diagnostic,
@@ -73,10 +73,12 @@ export interface ProjectedProgram extends Program {
   projector: Projector;
 }
 
-export function isProjectedProgram(
-  program: Program | ProjectedProgram
-): program is ProjectedProgram {
-  return "projector" in program;
+export function projectProgram(
+  program: Program,
+  projections: ProjectionApplication[],
+  startNode?: Type
+): ProjectedProgram {
+  return createProjector(program, projections, startNode);
 }
 
 export interface Program {
@@ -132,144 +134,9 @@ interface Validator {
   callback: (program: Program) => void | Promise<void>;
 }
 
-class StateMap extends Map<undefined | Projector, Map<Type, unknown>> {}
-class StateSet extends Map<undefined | Projector, Set<Type>> {}
-
-class StateMapView<V> implements Map<Type, V> {
-  public constructor(
-    private state: StateMap,
-    private projector?: Projector
-  ) {}
-
-  has(t: Type) {
-    return this.dispatch(t)?.has(t) ?? false;
-  }
-
-  set(t: Type, v: any) {
-    this.dispatch(t).set(t, v);
-    return this;
-  }
-
-  get(t: Type) {
-    return this.dispatch(t).get(t);
-  }
-
-  delete(t: Type) {
-    return this.dispatch(t).delete(t);
-  }
-
-  forEach(cb: (value: V, key: Type, map: Map<Type, V>) => void, thisArg?: any) {
-    this.dispatch().forEach(cb, thisArg);
-    return this;
-  }
-
-  get size() {
-    return this.dispatch().size;
-  }
-
-  clear() {
-    return this.dispatch().clear();
-  }
-
-  entries() {
-    return this.dispatch().entries();
-  }
-
-  values() {
-    return this.dispatch().values();
-  }
-
-  keys() {
-    return this.dispatch().keys();
-  }
-
-  [Symbol.iterator]() {
-    return this.entries();
-  }
-
-  [Symbol.toStringTag] = "StateMap";
-
-  dispatch(keyType?: Type): Map<Type, V> {
-    const key = keyType ? keyType.projector : this.projector;
-    if (!this.state.has(key)) {
-      this.state.set(key, new Map());
-    }
-
-    return this.state.get(key)! as any;
-  }
-}
-
-class StateSetView implements Set<Type> {
-  public constructor(
-    private state: StateSet,
-    private projector?: Projector
-  ) {}
-
-  has(t: Type) {
-    return this.dispatch(t)?.has(t) ?? false;
-  }
-
-  add(t: Type) {
-    this.dispatch(t).add(t);
-    return this;
-  }
-
-  delete(t: Type) {
-    return this.dispatch(t).delete(t);
-  }
-
-  forEach(cb: (value: Type, value2: Type, set: Set<Type>) => void, thisArg?: any) {
-    this.dispatch().forEach(cb, thisArg);
-    return this;
-  }
-
-  get size() {
-    return this.dispatch().size;
-  }
-
-  clear() {
-    return this.dispatch().clear();
-  }
-
-  values() {
-    return this.dispatch().values();
-  }
-
-  keys() {
-    return this.dispatch().keys();
-  }
-
-  entries() {
-    return this.dispatch().entries();
-  }
-
-  [Symbol.iterator]() {
-    return this.values();
-  }
-
-  [Symbol.toStringTag] = "StateSet";
-
-  dispatch(keyType?: Type): Set<Type> {
-    const key = keyType ? keyType.projector : this.projector;
-    if (!this.state.has(key)) {
-      this.state.set(key, new Set());
-    }
-
-    return this.state.get(key)!;
-  }
-}
-
 interface TypeSpecLibraryReference {
   path: string;
   manifest: NodePackage;
-}
-
-export function projectProgram(
-  program: Program,
-  projections: ProjectionApplication[],
-  startNode?: Type
-): ProjectedProgram {
-  return createProjector(program, projections, startNode);
 }
 
 export async function compile(
@@ -830,6 +697,7 @@ export async function compile(
    * @param emitter Emitter ref to run
    */
   async function runEmitter(emitter: EmitterRef) {
+    const { createAssetEmitter } = await import("../emitter-framework/asset-emitter.js");
     const context: EmitContext<any> = {
       program,
       emitterOutputDir: emitter.emitterOutputDir,
@@ -1226,36 +1094,6 @@ export async function compile(
 
     return program.checker.resolveTypeReference(node);
   }
-}
-
-export function createStateAccessors(
-  stateMaps: Map<symbol, StateMap>,
-  stateSets: Map<symbol, StateSet>,
-  projector?: Projector
-) {
-  function stateMap<T>(key: symbol): StateMapView<T> {
-    let m = stateMaps.get(key);
-
-    if (!m) {
-      m = new StateMap();
-      stateMaps.set(key, m);
-    }
-
-    return new StateMapView(m, projector);
-  }
-
-  function stateSet(key: symbol): StateSetView {
-    let s = stateSets.get(key);
-
-    if (!s) {
-      s = new StateSet();
-      stateSets.set(key, s);
-    }
-
-    return new StateSetView(s, projector);
-  }
-
-  return { stateMap, stateSet };
 }
 
 /**
