+++ packages/compiler/src/core/projected-program.ts
@@ -1,0 +1,7 @@
+import type { Program, ProjectedProgram } from "./index.js";
+
+export function isProjectedProgram(
+  program: Program | ProjectedProgram
+): program is ProjectedProgram {
+  return "projector" in program;
+}
