+++ packages/compiler/src/core/projector.ts
@@ -1,9 +1,11 @@
 import { createRekeyableMap, mutate } from "../utils/misc.js";
 import { finishTypeForProgram } from "./checker.js";
 import { compilerAssert } from "./diagnostics.js";
-import { Program, ProjectedProgram, createStateAccessors, isProjectedProgram } from "./program.js";
+import type { Program, ProjectedProgram } from "./program.js";
+import { isProjectedProgram } from "./projected-program.js";
+import { createStateAccessors } from "./state-accessors.js";
 import { getParentTemplateNode, isNeverType, isTemplateInstance, isValue } from "./type-utils.js";
-import {
+import type {
   DecoratorApplication,
   DecoratorArgument,
   Enum,
